02.03

+++ taejun.py
@@ -372,7 +372,7 @@ async def on_message(message):
             if "초벽" in message.content and "섹시" in message.content:
                 print(message.content)
                 await message.delete()
-            if "초벽오빠" in message.content:
+            if "초벽오빠" in message.content and "초벽 오빠":
                 print(message.content)
                 await message.delete()
             if message.channel.id == 926118022245142538 and (message.author.id != 263662225309433857 and message.author.id != 903288998577983530 and message.author.id != 397084939897667584):
